change dump - about to get rid of benching
rationale: it makes it very complicated to handle the player's current
position, and I feel that it does not meaningfully contribute to the
representation or gameplay. I originally wanted it because I thought
it would be nice to have a property that the player's coordinate would
always be defined, and players would be able to individually take a
mid-game break without everyone needing to pause together. But the
complexity it introduces to the types and getters is just not worth
it. In fact, I think the utility to a player of being able to pause
at such a granular level would just be annoying. It's been an awesome
challenge to try to accomplish, and it hurts to let that effort go to
nothing, but I think it will hurt me way more to keep beating at it
like a dead horse.