revert beNiceTo system back to teamId system
Rationale: the beNiceTo system was way too complicated, and I think
anybody trying to play the game would think the same thing. It had
some nice benefits of asymmetrical / non-mutual nicity, but I now
seriously believe that its easier on everyone if that kind of thing
is just enforced outside the game implementation and design.