debug tslib missing import of private field polyfill
this took so long to figure out. Interesting note: It would have taken
significantly less time if I had faith to believe that the weird problem
I was experiencing with __classPrivateFieldGet/Set was not just bad
luck and just googled it so specifically. oh well.

on a more reflective note, web development is insane. I wanted to make
a comparison to development in C++ based on my experience with sudoku,
but I guess the context is not very comparable, and now that I think
about it, I also remember feeling that C++ development is a little
insane based on all its nuance and gotchas. I guess in that respect,
they are comparable and both very insane.