change online-play screen nav flow
Made the lobby screen to be before the setup screen. Also made
a secure field in TopLevel to indicate whether the client is the
group host. This also solves a design problem about how to make
sure the group host receives UserInfoChange events while setting
up the game, which is not easy to implement when the setup screen
goes before the lobby screen.