upgrade to webpack 5
I'm so early to the party :0
(webpack 5's first stable release came out just yesterday)

It's kind of cool to be at a point where I'm not just reading about the
histories of technologies and things that have preceded me for years. At
the same time, change is hard, and waiting for good things is also hard
in its own way.