remove PlayOnline.onPlayerLeave
It's just way too complicated to design a way to destroy Player
instances from a Game that is still in use. There is nothing to
structurally prevent anything with access to a Game from getting
a hold to a Player, so it would be very difficult to enforce
everything to release a Player that is leaving.

So, a Player that has left the game will just stay on the board
as if their operator was AFK.