fix bug: online go game to lobby to game
That was pretty frustrating. I improved a lot of my code
along the way, but one of the main problems was some socket.io
behaviour related to autoconnection that doesn't make much
sense to me. When closing a socket and then reusing it with
Manager.socket, the socket does not automatically connect even
if its autoconnect setting is set to true...