Use custom module resolution strategy with leading colons
This solution was chosen to be easy to implement in both a tsconfig and
in a webpack config. The goal was to find a solution where modules could
be imported from the base or client/server in a super terse format, and
where if both the base and a client/server module had identically named
subfolders, one could be chosen without ambiguity. I couldn't find a way
to configure webpack resolver aliases as being based on the issuer path,
so I had to implement a resolver plugin. I really wanted to find a way
to implement it under `compiler.resolve.plugins`, but I couldn't find
any good examples of how to make one in webpack 5.

Previously, I could only choose between base or client/server "root"
imports being terse. This allows me to be terse for both. The reason why
both start with a colon instead of only one of them is because it makes
it super easy in the resolver plugin implementation to differentiate
whether an import is intended to be resolved using node module resolution
logic or not.